test: move representative benchmarks to proper test package
Now that we are no longer using jhump/protoreflect for this, these tests
can live in the same package as the other benchmarks.

Also mark file descriptor json files as binary so that they no longer
show up in diffs or git grep.
