+++ go.mod
@@ -10,9 +10,7 @@ require (
 )
 
 require (
-	github.com/bufbuild/protocompile v0.4.0 // indirect
 	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/jhump/protoreflect v1.15.1 // indirect
 	golang.org/x/net v0.9.0 // indirect
 	golang.org/x/sys v0.7.0 // indirect
 	golang.org/x/text v0.9.0 // indirect
