+++ test/mergedregistry/testdata/simd-allfd-c4489d.json
