+++ test/mergedregistry/testdata/simd-globalfd-c4489d.json
